Revert "sombrero.sh can now be called from wherever"
This reverts commit f6e0530746a1e711cf45ea2114da91ff648a5c46.
The reason that commit was done in the first place
was to allow spack to install sombrero.sh anywhere
and make so that sombrero.sh could invoke the subprograms.

This, though, can be fixed
in the spack package directly
using the 'edit()' function
changing sombrero/ to a full path.
So we can keep the original, less versatile version,
just to be compatible with what Mosè already did.